fix: detach docker container class from the tests
