+++ justfile
@@ -20,15 +20,20 @@ help:
     @echo " check: Check your code before committing (pre-commit hooks)"
     @echo
 
-# This target installs the development dependencies.
-install-dev:
+# This target installs the project dependencies.
+install-deps:
     export PATH="$HOME/miniconda3/bin:$PATH" && \
-        . ${HOME}/miniconda3/etc/profile.d/conda.sh && \
-        conda init && \
-        conda deactivate && \
-        conda activate test-actions && \
-        pip install -r requirements.txt && \
-        pip install -r requirements-dev.txt
+    . ${HOME}/miniconda3/etc/profile.d/conda.sh && \
+    conda init && \
+    conda deactivate && \
+    conda activate test-actions && \
+    pip install -e packages/localtest-docker && \
+    pip install -r requirements.txt && \
+    pip install -r requirements-dev.txt
+
+# This target installs the development dependencies.
+install-dev: install-deps
+    @echo "Installing development dependencies..."
 
 # This target is used to build the project.
 build:
@@ -49,7 +54,7 @@ run-local:
     docker run -it -p 5000:5000 -e ENVIRONMENT=development -e 'DATABASE_URL=sqlite:////data/test_db' -v $pwd/data:/data -v $pwd/src:/app test-actions
 
 # Use this command to run pytest
-test:
+test: install-deps
     rm -rf /$pwd/data/test_db && \
         export DATABASE_URL="sqlite:///$pwd/data/test_db" && \
         export PYTHONPATH=$pwd/src && \
@@ -59,11 +64,10 @@ test:
         conda init && \
         conda deactivate && \
         conda activate test-actions && \
-        pip install -r requirements-dev.txt && \
         pytest --cov src
 
 # Use this command to run pytest in CI/CD pipeline and generate coverage report
-test-ci:
+test-ci: install-deps
     cd $pwd && \
         rm -rf data || true && \
         mkdir -p data
@@ -71,8 +75,6 @@ test-ci:
         export PYTHONPATH=$pwd/src && \
         export ENVIRONMENT=development && \
         export HOST=0.0.0.0 && \
-        pip install -r requirements-dev.txt && \
-        pip install -r requirements.txt && \
         pytest --cov=src --cov-report=xml --cov-config=tox.ini --cov-branch src --full-trace
 
 # Use this command to check your code before committing (pre-commit hooks)
