+++ packages/localtest-docker/localtest_docker.egg-info/SOURCES.txt
@@ -1,0 +1,9 @@
+README.md
+setup.py
+localtest_docker/__init__.py
+localtest_docker/docker.py
+localtest_docker.egg-info/PKG-INFO
+localtest_docker.egg-info/SOURCES.txt
+localtest_docker.egg-info/dependency_links.txt
+localtest_docker.egg-info/requires.txt
+localtest_docker.egg-info/top_level.txt
